+++ react_frontend/src/styles/base.css
@@ -1,0 +1,306 @@
+:root {
+  color-scheme: light;
+  --charcoal: #222831;
+  --charcoal-strong: #121820;
+  --charcoal-soft: #3a424c;
+  --orange: #e86f19;
+  --orange-strong: #b84f0f;
+  --gold: #f5a33a;
+  --paper: #f7f4ee;
+  --paper-warm: #fff8ef;
+  --surface: #ffffff;
+  --surface-soft: #fbfaf7;
+  --mist: #edf1f3;
+  --muted: #67717d;
+  --line: rgba(34, 40, 49, 0.14);
+  --line-strong: rgba(34, 40, 49, 0.24);
+  --success: #28745f;
+  --warning: #d9891d;
+  --danger: #b2472e;
+  --shadow: 0 18px 48px rgba(18, 24, 32, 0.14);
+  --shadow-tight: 0 10px 26px rgba(18, 24, 32, 0.12);
+  --radius: 8px;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+html {
+  min-height: 100%;
+  background: var(--paper);
+}
+
+body {
+  margin: 0;
+  min-width: 320px;
+  background:
+    repeating-linear-gradient(90deg, rgba(34, 40, 49, 0.035) 0 1px, transparent 1px 72px),
+    repeating-linear-gradient(0deg, rgba(34, 40, 49, 0.028) 0 1px, transparent 1px 72px),
+    linear-gradient(135deg, #f7f4ee 0%, #eef1f3 47%, #fff4e8 100%);
+  background-attachment: fixed;
+  color: var(--charcoal);
+  font-family: Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+}
+
+body::selection {
+  background: rgba(232, 111, 25, 0.24);
+}
+
+#root {
+  min-height: 100vh;
+}
+
+button,
+input {
+  font: inherit;
+}
+
+button {
+  cursor: pointer;
+}
+
+button:disabled {
+  cursor: not-allowed;
+}
+
+button:focus-visible,
+input:focus-visible {
+  outline: 3px solid rgba(232, 111, 25, 0.32);
+  outline-offset: 2px;
+}
+
+.app-shell {
+  isolation: isolate;
+  min-height: 100vh;
+  padding: 24px;
+  position: relative;
+}
+
+.app-shell::before {
+  background:
+    linear-gradient(90deg, rgba(232, 111, 25, 0.12), transparent 36%),
+    linear-gradient(180deg, rgba(34, 40, 49, 0.08), transparent 22rem);
+  content: "";
+  inset: 0;
+  pointer-events: none;
+  position: fixed;
+  z-index: -1;
+}
+
+.topbar {
+  align-items: center;
+  background: rgba(255, 255, 255, 0.88);
+  backdrop-filter: blur(16px);
+  border: 1px solid rgba(255, 255, 255, 0.72);
+  border-radius: var(--radius);
+  box-shadow: var(--shadow-tight);
+  display: flex;
+  gap: 18px;
+  justify-content: space-between;
+  margin: 0 auto;
+  max-width: 1180px;
+  padding: 10px;
+  position: sticky;
+  top: 16px;
+  z-index: 10;
+}
+
+.brand {
+  align-items: center;
+  display: flex;
+  gap: 14px;
+  min-width: 0;
+}
+
+.brand-image {
+  background: var(--surface);
+  border: 1px solid rgba(34, 40, 49, 0.1);
+  border-radius: var(--radius);
+  box-shadow: inset 0 0 0 1px rgba(255, 255, 255, 0.72);
+  display: block;
+  height: 72px;
+  object-fit: cover;
+  object-position: center;
+  width: 220px;
+}
+
+.brand-summary {
+  border-left: 1px solid var(--line);
+  max-width: 300px;
+  min-width: 220px;
+  padding-left: 14px;
+}
+
+.brand-summary p {
+  color: var(--charcoal-soft);
+  font-size: 14px;
+  font-weight: 800;
+  line-height: 1.35;
+  margin: 0;
+}
+
+.topbar-actions {
+  align-items: center;
+  display: flex;
+  gap: 10px;
+  justify-content: flex-end;
+  min-width: 0;
+}
+
+.user-pill {
+  align-items: center;
+  background: #fff1e4;
+  border: 1px solid rgba(232, 111, 25, 0.24);
+  border-radius: 999px;
+  color: var(--orange-strong);
+  display: inline-flex;
+  flex: 0 0 auto;
+  font-size: 12px;
+  font-weight: 900;
+  min-height: 36px;
+  padding: 8px 12px;
+}
+
+.topbar nav {
+  background: rgba(34, 40, 49, 0.08);
+  border: 1px solid rgba(34, 40, 49, 0.08);
+  border-radius: var(--radius);
+  display: flex;
+  gap: 4px;
+  padding: 4px;
+}
+
+.topbar nav button {
+  background: transparent;
+  border: 0;
+  border-radius: 6px;
+  color: var(--charcoal-soft);
+  font-size: 14px;
+  font-weight: 850;
+  min-height: 38px;
+  padding: 9px 14px;
+  transition: background 160ms ease, color 160ms ease, transform 160ms ease;
+}
+
+.topbar nav button:hover {
+  background: rgba(255, 255, 255, 0.7);
+  color: var(--charcoal-strong);
+}
+
+.topbar nav button.active {
+  background: var(--charcoal);
+  box-shadow: 0 8px 18px rgba(18, 24, 32, 0.18);
+  color: var(--paper-warm);
+}
+
+.workspace-main {
+  display: grid;
+  gap: 22px;
+  margin: 24px auto 0;
+  max-width: 1180px;
+}
+
+.upload-card,
+.library-card,
+.login-card {
+  background:
+    linear-gradient(180deg, rgba(255, 255, 255, 0.98), rgba(251, 250, 247, 0.96)),
+    linear-gradient(90deg, rgba(232, 111, 25, 0.08), transparent 48%);
+  border: 1px solid var(--line);
+  border-radius: var(--radius);
+  border-top: 4px solid var(--orange);
+  box-shadow: var(--shadow);
+  overflow: hidden;
+  padding: 28px;
+  position: relative;
+}
+
+.upload-card::after,
+.library-card::after,
+.login-card::after {
+  background: linear-gradient(90deg, var(--orange), rgba(34, 40, 49, 0.14), transparent);
+  content: "";
+  height: 1px;
+  left: 0;
+  opacity: 0.7;
+  position: absolute;
+  right: 0;
+  top: 0;
+}
+
+.section-kicker {
+  color: var(--orange-strong);
+  font-size: 12px;
+  font-weight: 950;
+  letter-spacing: 0.16em;
+  margin-bottom: 12px;
+  text-transform: uppercase;
+}
+
+.upload-head,
+.library-head {
+  align-items: flex-start;
+  display: grid;
+  gap: 20px;
+  grid-template-columns: minmax(0, 1fr) auto;
+  margin-bottom: 24px;
+}
+
+.upload-head h1,
+.library-head h2,
+.login-card h1 {
+  color: var(--charcoal-strong);
+  font-size: 48px;
+  font-weight: 950;
+  letter-spacing: 0;
+  line-height: 1.03;
+  margin: 0;
+  max-width: 860px;
+}
+
+.library-head h2 {
+  font-size: 38px;
+}
+
+.login-card h1 {
+  font-size: 42px;
+  margin-bottom: 22px;
+  max-width: 520px;
+}
+
+.upload-head p,
+.library-head p,
+.resolution-toolbar p {
+  color: var(--muted);
+  font-size: 15px;
+  line-height: 1.6;
+  margin: 14px 0 0;
+  max-width: 700px;
+}
+
+.network-pill,
+.library-head > span {
+  align-items: center;
+  background: #fff1e4;
+  border: 1px solid rgba(232, 111, 25, 0.22);
+  border-radius: 999px;
+  color: var(--orange-strong);
+  display: inline-flex;
+  flex: 0 0 auto;
+  font-size: 13px;
+  font-weight: 900;
+  gap: 8px;
+  min-height: 38px;
+  padding: 9px 13px;
+}
+
+.network-pill::before,
+.library-head > span::before {
+  background: var(--orange);
+  border-radius: 50%;
+  box-shadow: 0 0 0 4px rgba(232, 111, 25, 0.12);
+  content: "";
+  height: 8px;
+  width: 8px;
+}
