+++ react_frontend/src/styles/player.css
@@ -1,0 +1,120 @@
+.player-shell {
+  align-items: center;
+  background:
+    linear-gradient(135deg, rgba(232, 111, 25, 0.12), transparent 42%),
+    #05070a;
+  border: 1px solid rgba(34, 40, 49, 0.32);
+  border-radius: var(--radius);
+  display: flex;
+  height: clamp(320px, 72vh, 760px);
+  height: clamp(320px, 72dvh, 760px);
+  justify-content: center;
+  margin-top: 16px;
+  overflow: hidden;
+  position: relative;
+}
+
+.player {
+  display: block;
+  height: 100%;
+  object-fit: contain;
+  width: 100%;
+}
+
+.player-error {
+  background: rgba(5, 7, 10, 0.84);
+  border: 1px solid rgba(255, 248, 239, 0.24);
+  border-radius: var(--radius);
+  bottom: 18px;
+  color: var(--paper-warm);
+  font-size: 13px;
+  font-weight: 850;
+  left: 18px;
+  max-width: min(420px, calc(100% - 36px));
+  padding: 10px 12px;
+  position: absolute;
+  z-index: 2;
+}
+
+.player-quality {
+  align-items: flex-end;
+  bottom: 18px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  opacity: 0;
+  pointer-events: none;
+  position: absolute;
+  right: 18px;
+  transform: translateY(6px);
+  transition: opacity 160ms ease, transform 160ms ease;
+  z-index: 2;
+}
+
+.player-shell.controls-active .player-quality,
+.player-quality.open {
+  opacity: 1;
+  pointer-events: auto;
+  transform: translateY(0);
+}
+
+.quality-toggle {
+  align-items: center;
+  background: rgba(5, 7, 10, 0.78);
+  border: 1px solid rgba(255, 248, 239, 0.22);
+  border-radius: 999px;
+  color: var(--paper-warm);
+  display: inline-flex;
+  font-size: 13px;
+  font-weight: 900;
+  gap: 7px;
+  min-height: 38px;
+  padding: 8px 12px;
+}
+
+.quality-toggle.active,
+.quality-toggle:hover {
+  background: rgba(255, 248, 239, 0.94);
+  color: var(--charcoal);
+}
+
+.gear-icon {
+  align-items: center;
+  display: inline-flex;
+  height: 16px;
+  justify-content: center;
+  line-height: 1;
+  width: 16px;
+}
+
+.gear-icon::before {
+  content: "\2699";
+  font-size: 16px;
+}
+
+.quality-menu {
+  background: rgba(5, 7, 10, 0.9);
+  border: 1px solid rgba(255, 248, 239, 0.2);
+  border-radius: var(--radius);
+  box-shadow: 0 18px 40px rgba(0, 0, 0, 0.36);
+  display: grid;
+  gap: 4px;
+  min-width: 112px;
+  padding: 6px;
+}
+
+.quality-menu button {
+  background: transparent;
+  border: 0;
+  border-radius: 6px;
+  color: var(--paper-warm);
+  font-size: 13px;
+  font-weight: 850;
+  padding: 8px 10px;
+  text-align: left;
+}
+
+.quality-menu button.active,
+.quality-menu button:hover {
+  background: rgba(255, 248, 239, 0.16);
+}
